fix: Resolve log() shadowing by global EchoelLogger and CGFloat type mismatch (#257)
- Qualify math log() calls with Darwin.log() across 4 files to avoid
  shadowing by global `let log = EchoelLogger.shared`
- Fix Float→CGFloat assignment in VJLaserControlView KeystoneKnob drag

+++ Sources/Echoelmusic/Views/VJLaserControlView.swift
@@ -1193,8 +1193,8 @@ struct KeystoneKnob: View {
         .gesture(
             DragGesture()
                 .onChanged { drag in
-                    x = Float(drag.translation.width / 50).clamped(to: -1...1)
-                    y = Float(drag.translation.height / 50).clamped(to: -1...1)
+                    x = CGFloat(drag.translation.width / 50).clamped(to: -1...1)
+                    y = CGFloat(drag.translation.height / 50).clamped(to: -1...1)
                 }
         )
     }
